Simplify service mod.
