+++ tunet-service/src/main.rs
@@ -1,17 +1,15 @@
 cfg_if::cfg_if! {
     if #[cfg(target_os = "windows")] {
         #[path = "windows/mod.rs"]
-        mod platform;
+        mod service;
     } else if #[cfg(target_os = "macos")] {
         #[path = "mac/mod.rs"]
-        mod platform;
+        mod service;
     } else {
         #[path ="stub/mod.rs"]
-        mod platform;
+        mod service;
     }
 }
-
-use platform::*;
 
 mod elevator;
 mod notification;
