+++ tunet-service/src/stub/mod.rs
@@ -1,15 +1,13 @@
-pub mod service {
-    use tunet_helper::{anyhow, Result};
+use tunet_helper::{anyhow, Result};
 
-    pub fn register(_interval: Option<humantime::Duration>) -> Result<()> {
-        Err(anyhow!("不支持的命令"))
-    }
+pub fn register(_interval: Option<humantime::Duration>) -> Result<()> {
+    Err(anyhow!("不支持的命令"))
+}
 
-    pub fn unregister() -> Result<()> {
-        Err(anyhow!("不支持的命令"))
-    }
+pub fn unregister() -> Result<()> {
+    Err(anyhow!("不支持的命令"))
+}
 
-    pub fn start(_interval: Option<humantime::Duration>) -> Result<()> {
-        Err(anyhow!("不支持的命令"))
-    }
+pub fn start(_interval: Option<humantime::Duration>) -> Result<()> {
+    Err(anyhow!("不支持的命令"))
 }
